Renamed loading command files
