+++ lib/index.js
@@ -32,11 +32,12 @@ client.defaultGuildConfig = {
   prefix: client.config.prefix,
 };
 
-client.loadSingleCommand = require('./loadSingleCommand');
-client.loadCommands = require('./loadCommands');
+client.loadSingleCommand = require('./loadOneCommand');
+client.loadCommands = require('./loadAllCommands');
 client.loadCommands(client)
   .then(() => {
-    logger.info(`Loaded ${client.commands.size} command(s), with ${client.aliases.size} aliases.`);
+    logger.info(`Loaded ${client.commands.size} command${client.commands.size === 1 ? '' : 's'}, ` +
+      `with ${client.aliases.size} aliases.`);
   })
   .catch(err => {
     logger.error('Error loading commands', err);
@@ -167,6 +168,8 @@ function prefixedCommands(msg) {
   if (!cmd) {
     return;
   }
+  // Save the prefix used on this message (guild) so we can use it to build guild specific help commands
+  msg.prefix = conf.prefix;
   logger.debug('Command:', cmd, JSON.stringify(params));
   // Attempt to get command from Collection
   let command;
