+++ lib/loadAllCommands.js
