+++ lib/loadOneCommand.js
